+++ src/euroeval/__init__.py
@@ -86,6 +86,13 @@ os.environ["RAY_DISABLE_DOCKER_CPU_WARNING"] = "1"
 os.environ["VLLM_WORKER_MULTIPROC_METHOD"] = "spawn"
 
 
+# Avoid the "Unclosed client session" error when evaluating Ollama models with LiteLLM.
+# The error comes from the `aiohttp` package, and this environment variable forces the
+# use of `httpx` instead.
+# Link: https://github.com/BerriAI/litellm/issues/11657#issuecomment-3038984975
+os.environ["DISABLE_AIOHTTP_TRANSPORT"] = "True"
+
+
 # Use older version v0 of vLLM, as the newer one requires XGrammar as decoding backend,
 # but XGrammar does not support having a maximal amount of elements in lists
 os.environ["VLLM_USE_V1"] = "0"