+++ tests/test_data_models.py
@@ -20,7 +20,7 @@ def test_all_classes_are_dataclasses_or_pydantic_models() -> None:
         if not obj_name.startswith("_")
         and inspect.isclass(object=getattr(data_models, obj_name))
         and not hasattr(enums, obj_name)
-        and obj_name != "ScoreDict"
+        and obj_name not in {"ScoreDict", "Metric"}
     ]
     for obj in all_classes:
         obj_is_dataclass = hasattr(obj, "__dataclass_fields__")